backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
psycopg[binary]>=3.1,<4
PyJWT>=2.8.0
cryptography>=41.0.0
werkzeug>=2.3.0
